Fix crash when load more itens quickly
