+++ app/build.gradle
@@ -78,8 +78,6 @@ dependencies {
     compile 'com.github.bumptech.glide:glide:4.0.0-RC0'
     kapt 'com.github.bumptech.glide:compiler:4.0.0-RC0'
 
-    compile 'jp.wasabeef:recyclerview-animators:2.2.6'
-
     compile 'nz.bradcampbell:paperparcel:2.0.1'
     kapt 'nz.bradcampbell:paperparcel-compiler:2.0.1'
 }
